+++ src/OctoshiftCLI.Tests/bbs2gh/Commands/MigrateRepo/MigrateRepoCommandHandlerTests.cs
@@ -528,7 +528,6 @@ namespace OctoshiftCLI.Tests.BbsToGithub.Commands.MigrateRepo
                 BbsPassword = BBS_PASSWORD,
                 BbsProject = BBS_PROJECT,
                 BbsRepo = BBS_REPO,
-                Wait = true
             };
 
             // Assert