+++ src/OctoshiftCLI.Tests/gei/Commands/MigrateRepo/MigrateRepoCommandArgsTests.cs
@@ -9,9 +9,6 @@ namespace OctoshiftCLI.Tests.GithubEnterpriseImporter.Commands.MigrateRepo
     public class MigrateRepoCommandArgsTests
     {
         private readonly Mock<OctoLogger> _mockOctoLogger = TestHelpers.CreateMock<OctoLogger>();
-
-        private const string GHES_API_URL = "https://myghes/api/v3";
-        private const string AZURE_CONNECTION_STRING = "DefaultEndpointsProtocol=https;AccountName=myaccount;AccountKey=mykey;EndpointSuffix=core.windows.net";
         private const string SOURCE_ORG = "foo-source-org";
         private const string SOURCE_REPO = "foo-repo-source";
         private const string TARGET_ORG = "foo-target-org";
@@ -27,7 +24,6 @@ namespace OctoshiftCLI.Tests.GithubEnterpriseImporter.Commands.MigrateRepo
                 GithubSourceOrg = SOURCE_ORG,
                 SourceRepo = SOURCE_REPO,
                 GithubTargetOrg = TARGET_ORG,
-                Wait = true
             };
 
             args.Validate(_mockOctoLogger.Object);
@@ -105,69 +101,6 @@ namespace OctoshiftCLI.Tests.GithubEnterpriseImporter.Commands.MigrateRepo
                 .Should()
                 .ThrowExactly<OctoshiftCliException>()
                 .WithMessage("*--keep-archive*");
-        }
-
-        [Fact]
-        public void Validates_Wait_And_QueueOnly_Not_Passed_Together()
-        {
-            var args = new MigrateRepoCommandArgs
-            {
-                GithubSourceOrg = SOURCE_ORG,
-                SourceRepo = SOURCE_REPO,
-                GithubTargetOrg = TARGET_ORG,
-                TargetRepo = TARGET_REPO,
-                GhesApiUrl = GHES_API_URL,
-                AzureStorageConnectionString = AZURE_CONNECTION_STRING,
-                Wait = true,
-                KeepArchive = true,
-                QueueOnly = true,
-            };
-
-            FluentActions.Invoking(() => args.Validate(_mockOctoLogger.Object))
-                               .Should()
-                               .ThrowExactly<OctoshiftCliException>()
-                               .WithMessage("*wait*");
-        }
-
-        [Fact]
-        public void Wait_Flag_Shows_Warning()
-        {
-            var args = new MigrateRepoCommandArgs
-            {
-                GithubSourceOrg = SOURCE_ORG,
-                SourceRepo = SOURCE_REPO,
-                GithubTargetOrg = TARGET_ORG,
-                TargetRepo = TARGET_REPO,
-                GhesApiUrl = GHES_API_URL,
-                AzureStorageConnectionString = AZURE_CONNECTION_STRING,
-                Wait = true,
-                KeepArchive = true,
-            };
-
-            args.Validate(_mockOctoLogger.Object);
-
-            _mockOctoLogger.Verify(x => x.LogWarning(It.Is<string>(x => x.ToLower().Contains("wait"))));
-        }
-
-        [Fact]
-        public void No_Wait_And_No_Queue_Only_Flags_Shows_Warning()
-        {
-            var args = new MigrateRepoCommandArgs
-            {
-                GithubSourceOrg = SOURCE_ORG,
-                SourceRepo = SOURCE_REPO,
-                GithubTargetOrg = TARGET_ORG,
-                TargetRepo = TARGET_REPO,
-                GhesApiUrl = GHES_API_URL,
-                AzureStorageConnectionString = AZURE_CONNECTION_STRING,
-                Wait = false,
-                QueueOnly = false,
-                KeepArchive = true,
-            };
-
-            args.Validate(_mockOctoLogger.Object);
-
-            _mockOctoLogger.Verify(x => x.LogWarning(It.Is<string>(x => x.ToLower().Contains("wait"))));
         }
     }
 }